OM2-186 | Enable multiple region AWS services
We're having an issue to serve S3 files from `sa-east-1` to fulfillers
because our last AWS usage approach only allowed us to access one
region at a time, which seems to be a limitation from the Aws SDK
Bundle we used to use.

Now we've just got rid of that bundle and, instead, opted to use the
AWS SDK directly, configuring the resources as we need to.

Given that fact, now we MUST set the **global** configs on the config
file, using the following pattern:

```yaml
printi_aws:
  global:
    version: latest
    region: us-east-1
    credentials:
      key: "%env(AWS_KEY)%"
      secret: "%env(AWS_SECRET)%"
```

Also, we're now able to set the region for each resource we have:

```yaml
printi_aws:
  global: (...)
  s3:
    orders_bucket:
      name: alpha-upload-dev
      region: sa-east-1
```

If the region is not provided, the **global** will be used instead.
If no **global** region is set, `us-east-1` will be used by default.