Improved GetUrlInfo method to add region
- Improved S3::getUrlInfo() method to add the region to the capture
groups, allowing us to generate a presigned URL even for buckets we may
not have all the configs for.

- Also made some adjustments on what we do when we can't parse the URL:
now we throw an exception instead of letting the process fail almost
silently by having a notice that a key is not set, since we returned an
empty array, which doesn't help by any means the caller.

- Changed a bit AwsService, making the properties protected instead of
private, since it's an abstract class.

Those, specially the 1st and 2nd points, will allow us to interact with
"legacy" orders that are not on the new OM2 S3 bucket standard paths.